[v0.88][runtime] Split provider orchestration into smaller modules (Closes #1719)
